chore: add a docs report actions for PRs
