+++ package.json
@@ -31,6 +31,9 @@
     "dev:starter-studio": "yarn --cwd dev/starter-studio dev",
     "dev:strict-studio": "yarn --cwd dev/strict-studio dev",
     "dev:test-studio": "yarn --cwd dev/test-studio dev",
+    "docs:report": "node -r dotenv-flow/config -r esbuild-register scripts/docReport",
+    "docs:report:create": "node -r dotenv-flow/config -r esbuild-register scripts/docReportCreate",
+    "docs:report:cleanup": "node -r dotenv-flow/config -r esbuild-register scripts/docReportCleanup",
     "eslint": "eslint --ext=.cjs,.js,.jsx,.mjs,.ts,.tsx --quiet",
     "example:blog-studio": "yarn --cwd examples/blog-studio start",
     "example:clean-studio": "yarn --cwd examples/clean-studio start",
@@ -38,7 +41,6 @@
     "example:example-studio": "yarn --cwd dev/example-studio start",
     "example:movies-studio": "yarn --cwd examples/movies-studio start",
     "etl": "node -r dotenv-flow/config -r esbuild-register scripts/etl",
-    "docs": "node -r dotenv-flow/config -r esbuild-register scripts/docReport",
     "init": "lerna clean --yes && run-s bootstrap build",
     "lerna:clean": "lerna clean",
     "lint": "eslint --cache .",
