+++ perf/config/envVars.ts
@@ -6,6 +6,8 @@ type KnownEnvVar =
   | 'COMMIT_SHA'
   | 'BRANCH_DEPLOYMENT_URL'
   | 'SANITY_STUDIO_DATASET'
+  | 'DOCS_REPORT_DATASET'
+  | 'DOCS_REPORT_TOKEN'
 
 export function readEnv(name: KnownEnvVar): string {
   const val = findEnv(name)
