+++ scripts/utils/sanityIdify.ts
@@ -1,0 +1,7 @@
+import {deburr} from 'lodash'
+
+export function sanityIdify(input: string): string {
+  return deburr(input)
+    .replace(/[^a-zA-Z0-9_-]+/g, '_')
+    .replace(/^-/, '')
+}
